Bug Fixes in heatmaps.php  (#172)
I fixed a few bugs in heatmaps.php where the file name for the Heatmap &
Screenshot did not refer to the source video.
I also fixed a bug in the same file where the recording information on
the overlay on the video display would not change when the video was
changed. It would remain the same regardless of the video being viewed.
I broke this file fix into its one branch to expedite the code review
and approval.